+++ src/pages/Company/CompanyTimeAndSalaryProvider.js
@@ -26,9 +26,11 @@ import {
 import { paramsSelector, querySelector } from 'common/routing/selectors';
 import useCompanyName, { companyNameSelector } from './useCompanyName';
 import { useTopNJobTitles } from './useTopNJobTitles';
-import { pageFromQuerySelector } from 'selectors/routing/page';
+import {
+  queryFromQuerySelector,
+  pageFromQuerySelector,
+} from 'selectors/routing';
 import { useSearchTextFromQuery } from 'components/CompanyAndJobTitle/Searchbar';
-import { keywordFromQuerySelector } from 'selectors/routing/keyword';
 
 const useOverviewStatisticsBox = pageName => {
   const selector = useMemo(
@@ -167,7 +169,7 @@ CompanyTimeAndSalaryProvider.fetchData = ({
   const companyName = companyNameSelector(params);
   const query = querySelector(props);
   const page = pageFromQuerySelector(query);
-  const jobTitle = keywordFromQuerySelector(query) || undefined;
+  const jobTitle = queryFromQuerySelector(query) || undefined;
   const start = (page - 1) * PAGE_SIZE;
   const limit = PAGE_SIZE;
   const dispatchOverviewStatistics = dispatch(
