+++ src/pages/JobTitle/JobTitleIndexProvider.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { querySelector } from 'common/routing/selectors';
-import { pageFromQuerySelector } from 'selectors/routing/page';
+import { pageFromQuerySelector } from 'selectors/routing';
 import CompanyAndJobTitleIndexPage from 'components/CompanyAndJobTitle/IndexPage';
 import usePagination from 'components/CompanyAndJobTitle/IndexPage/usePagination';
 import { pageType as PAGE_TYPE, PAGE_SIZE } from 'constants/companyJobTitle';
