--- src/selectors/routing/page.js
@@ -1,8 +1,0 @@
-import { compose, when, always } from 'ramda';
-
-// page from ?p=xxx
-export const pageFromQuerySelector = compose(
-  when(Number.isNaN, always(1)),
-  p => parseInt(p, 10),
-  query => query.p,
-);
